Simplify API
Simplify the API so that functions in the chain are only responsible
for returning new props for the component that they care about. This
helps to abstract any side affects from previous functions in that
they set a prop that another function depends on. This way all functions
get the same properties and are expected to return new props or nothing.